requirements.txt
gevent
grpcio
protobuf
attrs
bcrypt
cffi
click
Flask
Flask-Bcrypt
Flask-Login
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
itsdangerous
Jinja2
MarkupSafe
pluggy
py
pycparser
pytest
six
SQLAlchemy
Werkzeug
WTForms
protobuf3-to-dict
simplejson